Verify the route owner authentification.

+++ apps/routes-api/src/app/services/routes.service.ts
@@ -2,16 +2,20 @@ import { Route, RouteModel } from "../models/route.model"
 import { NotifierService } from "./notifier.service"
 
 export class RoutesService {
-
     private notifierService: NotifierService
 
     constructor() {
         this.notifierService = new NotifierService()
     }
 
-    async removeRoute(id: string): Promise<Route> {
+    async removeRoute(id: string, loggedUserId: string): Promise<Route> {
+        const route = await this.getRouteById(id)
+        if (route.userId !== loggedUserId) {
+            throw new Error('You are not allowed to access this resource')
+        }        
         const result = await RouteModel
             .findByIdAndDelete(id)
+
         return result
     }
 
@@ -28,8 +32,17 @@ export class RoutesService {
             .exec()
     }
 
-    async updateRoute(putRoute: Route, id: string): Promise<Route> {
-        const result = await RouteModel.findByIdAndUpdate(id, putRoute)
+    async updateRoute(putRoute: Route, id: string, loggedUserId: string): Promise<Route> {
+        const route = await this.getRouteById(id)
+        if (route.userId !== loggedUserId) {
+            throw new Error('You are not allowed to access this resource')
+        }
+        const result = await RouteModel
+            .findByIdAndUpdate(id, putRoute)
         return result
+    }
+
+    async listRoutes(userId: string): Promise<Route[]> {
+        return RouteModel.find({ userId }).exec()
     }
 }
